play around with injecting macro results into cel context
